Update Customer 28-3-2022

+++ app/src/main/java/com/comcept/rmrscustomer/data_class/verifyInvoice/Discount.kt
@@ -5,8 +5,8 @@ import com.google.gson.annotations.SerializedName
 data class Discount(
 
     @SerializedName("AmountAfterDiscount") var AmountAfterDiscount: Double? = null,
-    @SerializedName("AmountBeforeDiscount") var AmountBeforeDiscount: Int? = null,
+    @SerializedName("AmountBeforeDiscount") var AmountBeforeDiscount: Float? = null,
     @SerializedName("DiscountName") var DiscountName: String? = null,
-    @SerializedName("DiscountPercentage") var DiscountPercentage: Int? = null,
-    @SerializedName("OtherDiscount") var OtherDiscount: Int? = null
+    @SerializedName("DiscountPercentage") var DiscountPercentage: Float? = null,
+    @SerializedName("OtherDiscount") var OtherDiscount: Float? = null
 )
